Change accelerometer calibration to 8 bit integers
16 bit is way to much for accelerometer calibration. Scale down to 8 bit
integers to support the linux evdev interface.

+++ driver/include/proto.h
@@ -312,9 +312,9 @@ struct wii_proto_cc_led {
 };
 
 struct wii_proto_cc_acalib {
-	int16_t x;
-	int16_t y;
-	int16_t z;
+	int8_t x;
+	int8_t y;
+	int8_t z;
 };
 
 /*
